✨ improve(@roots/bud): cli
