+++ sources/@roots/bud/src/cli/helpers/detectPackageManager.tsx
@@ -1,12 +1,63 @@
 import type {Bud} from '@roots/bud-framework'
 
+/**
+ * Has lockfile
+ */
+export const hasLock = (bud: Bud, lockfile: string) =>
+  bud.context.config?.[lockfile]
+
+/**
+ * Concretely specified
+ */
+export const isSpecified = (bud: Bud, packageManagerIdent: string) =>
+  bud.context.manifest?.packageManager?.includes(packageManagerIdent)
+
+/**
+ * NPM specified
+ */
+/**
+ * Seems to be yarn
+ */
 export const isYarn = (bud: Bud) =>
-  bud.context.config?.[`yarn.lock`] ||
-  bud.context.manifest?.packageManager?.includes(`yarn`)
+  hasLock(bud, `yarn.lock`) || isSpecified(bud, `yarn`)
 
+/**
+ * Seems to be npm
+ */
 export const isNpm = (bud: Bud) =>
-  bud.context.config?.[`package-lock.json`] ||
-  bud.context.manifest?.packageManager?.includes(`npm`)
+  hasLock(bud, `package-lock.json`) || isSpecified(bud, `npm`)
 
-export const detectPackageManager = (bud: Bud): `yarn` | `npm` | false =>
-  isYarn(bud) ? `yarn` : isNpm(bud) ? `npm` : false
+/**
+ * Ambigrous package manager
+ */
+export const isAmbiguous = (bud: Bud) =>
+  !isSpecified(bud, `yarn`) &&
+  !isSpecified(bud, `npm`) &&
+  isYarn(bud) &&
+  isNpm(bud)
+
+/**
+ * Indeterminate package manager
+ */
+export const isUnknown = (bud: Bud) => !isYarn(bud) && !isNpm(bud)
+
+/**
+ * Detect package manager
+ */
+export const detectPackageManager = (bud: Bud): `yarn` | `npm` | false => {
+  if (isUnknown(bud)) {
+    bud.warn(
+      `Package manager could not be determined. bud.js cli supports yarn and npm.`,
+    )
+    return false
+  }
+
+  if (isAmbiguous(bud)) {
+    bud.warn(
+      `Package manager is ambiguous. Default package manager is yarn.`,
+    )
+    return `yarn`
+  }
+
+  return isNpm(bud) ? `npm` : `yarn`
+}
